fix(nginx): CORS headers rewrited all headers (#51)
Apparently NGINX will take only the headers from the deepest
context that he finds. In this case, the deepest one was the
`if` condition on cors.conf. For this reason, all the
other headers were ignored.

This fix uses variables instead, so that all the `add_header`
directives stay on the same level, allowing us to not re-write
those headers.